database app/writer/src/requirements.txt
psycopg2-binary
google-cloud-secret-manager
google-cloud-logging